make TUI work after engine refactor (#146)
* make TUI work after engine refactor

* More straightforward exit handling

Use `on_unmount` method to shutdown gracefully, instead of overriding `action_quit`.  `on_unmount` is called whenever the app exits (even on an exception) and handles more shutdown cases. It is the best place to put code which must run on exit.  Also, overriding `action_quit` leads to type checking errors, because the method signature is not the same as in the super class.

* fix typing errors

* refactor: Use explicit parameters in TypingEffect constructor

Refactors the __init__ method of the TypingEffect widget to accept explicit parameters for the underlying Static widget instead of using generic *args and **kwargs. This improves type hinting and code clarity.

---------

Co-authored-by: David Noel Ng <dnhkng@gmail.com>